converted to uint16_t for dealing with overflows

+++ include/dct/twostep_efficient.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void dct_twostep_efficient(const uint8_t data_in[8][8], uint8_t data_out[8][8]);
+void dct_twostep_efficient(const uint8_t data_in[8][8], uint16_t data_out[8][8]);
 
 #endif // SENG440_TWOSTEP_EFFICIENT_H
