+++ src/dct/loeffler_float.c
@@ -84,7 +84,7 @@ void dct_1d(const double data_in[8], double data_out[8])
     data_out[1] = stage3out[7] + stage3out[4];
 }
 
-void dct_loeffler_float(const uint8_t data_in[8][8], uint8_t data_out[8][8])
+void dct_loeffler_float(const uint8_t data_in[8][8], uint16_t data_out[8][8])
 {
     // [X] = [C] * [x] * [C]^T
     double tmp_io[8][8];
@@ -111,7 +111,7 @@ void dct_loeffler_float(const uint8_t data_in[8][8], uint8_t data_out[8][8])
     for (int i = 0; i < 8; ++i) {
         for (int j = 0; j < 8; ++j) {
             // TODO: fix values outside [-128, 127]
-            data_out[i][j] = (uint8_t) tmp_io[i][j];
+            data_out[i][j] = (uint16_t) tmp_io[i][j];
         }
     }
 }
