+++ src/dct/twostep_slow.c
@@ -55,7 +55,7 @@ void square_matrix_multiply(const double lhs[8][8], const double rhs[8][8], doub
 
 // input: 8x8 array, output: 8x8 array.
 // uses the row-column separation approach from slides 8-10.
-void dct_twostep_slow(const uint8_t data_in[8][8], uint8_t data_out[8][8])
+void dct_twostep_slow(const uint8_t data_in[8][8], uint16_t data_out[8][8])
 {
     // [X] = [C] * [x] * [C]^T
     double tmp_io[8][8];
@@ -73,7 +73,7 @@ void dct_twostep_slow(const uint8_t data_in[8][8], uint8_t data_out[8][8])
     for (int i = 0; i < 8; ++i) {
         for (int j = 0; j < 8; ++j) {
             // TODO: fix values outside [-128, 127]
-            data_out[i][j] = (uint8_t) tmp_io[i][j];
+            data_out[i][j] = (uint16_t) tmp_io[i][j];
         }
     }
 }
